+++ f-list.h
@@ -131,6 +131,7 @@ typedef struct FListFriends_ FListFriends;
 #define FLIST_GLOBAL_TIMEOUT                    "TMO"
 #define FLIST_REWARD                            "RWD"
 #define FLIST_REQUEST_PENDING_REPORTS           "PCR"
+#define FLIST_REQUEST_UPTIME                    "UPT"
 
 #define LOGIN_TIMEOUT 5000