+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## ?.?.?
 ### Added
+- #55: New /uptime command : get server start date and some statistics
 ### Fixed
 - #130: Do not automatically try to re-login if kicked out from another location
 - The user now gets a helpful error message instead of a popup when trying to send a message beyond server size limits